intelligence_processor/requirements.txt
azure-functions
openai
sqlalchemy
psycopg2-binary
azure-search-documents==11.4.0
azure-storage-blob